Modification SalesManager int to long

+++ src/Main.java
@@ -1,10 +1,10 @@
 public class Main {
 
     public static void main(String[] args) {
-        int[] arr = {100, 200, 300};
+        long[] arr = {100, 200, 300};
         SalesManager salesManager = new SalesManager(arr);
 
-        int max = salesManager.max();
+        long max = salesManager.max();
 
         System.out.println(max);
